+++ data-prepper-plugins/kafka-plugins/src/integrationTest/java/org/opensearch/dataprepper/plugins/kafka/source/KafkaSourceMultipleAuthTypeIT.java
@@ -23,9 +23,9 @@ import org.opensearch.dataprepper.model.configuration.PipelineDescription;
 import org.opensearch.dataprepper.model.event.Event;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.plugins.kafka.configuration.AuthConfig;
+import org.opensearch.dataprepper.plugins.kafka.configuration.ConsumerTopicConfig;
 import org.opensearch.dataprepper.plugins.kafka.configuration.EncryptionConfig;
 import org.opensearch.dataprepper.plugins.kafka.configuration.EncryptionType;
-import org.opensearch.dataprepper.plugins.kafka.configuration.KafkaSourceConfig;
 import org.opensearch.dataprepper.plugins.kafka.configuration.PlainTextAuthConfig;
 import org.opensearch.dataprepper.plugins.kafka.configuration.TopicConfig;
 import org.opensearch.dataprepper.plugins.kafka.extension.KafkaClusterConfigSupplier;
@@ -67,7 +67,7 @@ public class KafkaSourceMultipleAuthTypeIT {
     private List<TopicConfig> topicList;
 
     @Mock
-    private TopicConfig plainTextTopic;
+    private ConsumerTopicConfig plainTextTopic;
 
     @Mock
     private AuthConfig authConfig;
@@ -128,7 +128,7 @@ public class KafkaSourceMultipleAuthTypeIT {
 
         final String testGroup = "TestGroup_"+RandomStringUtils.randomAlphabetic(6);
         final String testTopic = "TestTopic_"+RandomStringUtils.randomAlphabetic(5);
-        plainTextTopic = mock(TopicConfig.class);
+        plainTextTopic = mock(ConsumerTopicConfig.class);
         when(plainTextTopic.getName()).thenReturn(testTopic);
         when(plainTextTopic.getGroupId()).thenReturn(testGroup);
         when(plainTextTopic.getWorkers()).thenReturn(1);
@@ -154,7 +154,7 @@ public class KafkaSourceMultipleAuthTypeIT {
         final int numRecords = 1;
         when(encryptionConfig.getType()).thenReturn(EncryptionType.NONE);
         when(plainTextTopic.getConsumerMaxPollRecords()).thenReturn(numRecords);
-        when(sourceConfig.getTopics()).thenReturn(List.of(plainTextTopic));
+        when(sourceConfig.getTopics()).thenReturn((List) List.of(plainTextTopic));
         when(sourceConfig.getAuthConfig()).thenReturn(null);
         kafkaSource = createObjectUnderTest();
         
@@ -203,7 +203,7 @@ public class KafkaSourceMultipleAuthTypeIT {
         saslAuthConfig = mock(AuthConfig.SaslAuthConfig.class);
         when(encryptionConfig.getType()).thenReturn(EncryptionType.NONE);
         when(plainTextTopic.getConsumerMaxPollRecords()).thenReturn(numRecords);
-        when(sourceConfig.getTopics()).thenReturn(List.of(plainTextTopic));
+        when(sourceConfig.getTopics()).thenReturn((List) List.of(plainTextTopic));
         plainTextAuthConfig = mock(PlainTextAuthConfig.class);
         when(plainTextAuthConfig.getUsername()).thenReturn(kafkaUsername);
         when(plainTextAuthConfig.getPassword()).thenReturn(kafkaPassword);
@@ -262,7 +262,7 @@ public class KafkaSourceMultipleAuthTypeIT {
         when(encryptionConfig.getType()).thenReturn(EncryptionType.SSL);
         when(plainTextTopic.getConsumerMaxPollRecords()).thenReturn(numRecords);
         when(sourceConfig.getBootstrapServers()).thenReturn(Collections.singletonList(sslBootstrapServers));
-        when(sourceConfig.getTopics()).thenReturn(List.of(plainTextTopic));
+        when(sourceConfig.getTopics()).thenReturn((List) List.of(plainTextTopic));
         kafkaSource = createObjectUnderTest();
         
         Properties props = new Properties();
@@ -318,7 +318,7 @@ public class KafkaSourceMultipleAuthTypeIT {
         when(encryptionConfig.getType()).thenReturn(EncryptionType.SSL);
         when(plainTextTopic.getConsumerMaxPollRecords()).thenReturn(numRecords);
         when(sourceConfig.getBootstrapServers()).thenReturn(Collections.singletonList(saslsslBootstrapServers));
-        when(sourceConfig.getTopics()).thenReturn(List.of(plainTextTopic));
+        when(sourceConfig.getTopics()).thenReturn((List) List.of(plainTextTopic));
         kafkaSource = createObjectUnderTest();
         
         Properties props = new Properties();
