+++ data-prepper-plugins/kafka-plugins/src/main/java/org/opensearch/dataprepper/plugins/kafka/configuration/KafkaConsumerConfig.java
@@ -1,14 +1,18 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package org.opensearch.dataprepper.plugins.kafka.configuration;
 
 import java.util.List;
 
 public interface KafkaConsumerConfig extends KafkaConnectionConfig {
-
     String getClientDnsLookup();
 
     boolean getAcknowledgementsEnabled();
     
     SchemaConfig getSchemaConfig();
 
-    List<TopicConfig> getTopics();
+    List<? extends TopicConfig> getTopics();
 }
